requirements.txt
python-dotenv==1.0.1
instagrapi==1.18.0
openai==1.14.3
Pillow==9.5.0
requests==2.31.0
